add error checking to wp_build

+++ lib/build.php
@@ -3,7 +3,9 @@ function build_and_release($slug, $rootPath, $deployKey, $url){
   $zip = new ZipArchive();
   $filename = $rootPath . '/' . $slug . '.zip';
 
-  unlink($filename);
+  if(file_exists($filename)){
+    unlink($filename);
+  }
 
   if($zip->open($filename, ZipArchive::CREATE) !== true){
     exit("Could not create {$filename}");
@@ -35,6 +37,8 @@ function build_and_release($slug, $rootPath, $deployKey, $url){
 
   $zip->close();
 
+  echo('Zip Built' . PHP_EOL);
+
   $postOpts = array(
     'action' => 'wup_release',
     'deployKey' => $deployKey,
@@ -55,6 +59,12 @@ function build_and_release($slug, $rootPath, $deployKey, $url){
   $details = json_decode($result);
 
   curl_close($request);
+
+  if(!isset($details->error) && !isset($details->success)){
+    var_dump($details);
+    echo('Something went wrong submitting to ' . $url . PHP_EOL);
+    return;
+  }
 
   if(isset($details->error)){
     echo($details->error);
